requirements.txt
pytest
requests
selenium
allure-pytest
pytest-xdist